hostapd: Accept interface names as a command line parameter
This introduces an optinal command line parameter '-i' to override the
value of the 'interface' attribute in hostapd.conf files. This change
enables the reuse of the configuration files for the concurrent
instances of hostapd. An ability to dynamically assign the interface
names simplifies the usages of hostapd service in the automated
emulations of the wireless environments.

Signed-off-by: Tim Kourt <tim.a.kourt@linux.intel.com>